backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
playwright==1.40.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0
numpy==1.26.2
langchain==0.0.340
openai==1.3.5
httpx==0.25.1
boto3==1.34.0  # For S3 storage